Continue audio agent integration
