+++ src/main/scala/it/unibo/pps1920/motoscala/controller/SoundController.scala
@@ -1,5 +1,14 @@
 package it.unibo.pps1920.motoscala.controller
 
+import it.unibo.pps1920.motoscala.controller.managers.audio.{Clips, Media}
+
 trait SoundController {
+  def playSoundTrack(media: Media): Unit
+  def playEffect(media: Clips): Unit
+  def setVolumeSoundTrack(value: Double): Unit
+  def setVolumeEffect(value: Double): Unit
+  def stopMusic(): Unit
+  def restartMusic(): Unit
+  def pauseMusic(): Unit
 
 }
