+++ src/main/scala/it/unibo/pps1920/motoscala/controller/managers/audio/MediaEvent.scala
@@ -1,5 +1,6 @@
 package it.unibo.pps1920.motoscala.controller.managers.audio
 
 trait MediaEvent {
-
+  def handle(action: SoundAgentLogic => Unit): Unit
 }
+
